Refactor to abstract away Toolbar class
