+++ apps/equirect-buttons.js
@@ -4,9 +4,9 @@ import { XRControllerModelFactory } from "three/examples/jsm/webxr/XRControllerM
 
 import panoVideo from "../media/pano.mp4";
 import buttonClickSound from "../media/audio/button-click.mp3";
-import { CanvasUI } from "../util/CanvasUI";
 import { WebGLRenderer } from "../util/WebGLRenderer";
 import { VRButton } from "../util/webxr/VRButton";
+import Toolbar from "../util/Toolbar";
 
 class App {
     constructor(videoIn = panoVideo) {
@@ -26,23 +26,19 @@ class App {
         // Create Orbit Controls
         this.controls = this.createOrbitControls();
 
+        // Create Video
+        this.video = this.createVideo(videoIn);
+
         // Track which objects are hit
         this.raycaster = new THREE.Raycaster();
 
-        // Create Canvas UI
-        this.ui = this.createUI();
-
-        // Create Progress Bar
-        this.progressBar = this.createProgressBar();
-
         // Create Toolbar Group
-        this.toolbarGroup = this.createToolbarGroup();
+        // prettier-ignore
+        this.toolbar = new Toolbar(this.renderer, this.video, true);
+        this.toolbarGroup = this.toolbar.toolbar;
 
         // Hide the toolbar initially
         this.scene.userData.isToolbarVisible = false;
-
-        // Create Video
-        this.video = this.createVideo(videoIn);
 
         this.setupVR();
 
@@ -58,12 +54,12 @@ class App {
         const xr = this.renderer.xr;
         const session = xr.getSession();
 
-        if (xr.isPresenting) {
-            this.ui.update();
+        if (xr.isPresenting && this.toolbarGroup) {
+            this.toolbar.updateUI();
         }
 
-        if (this.video) {
-            this.updateProgressBar();
+        if (this.toolbar.video) {
+            this.toolbar.updateProgressBar();
         }
 
         if (
@@ -203,159 +199,6 @@ class App {
     }
 
     /**
-     * Creates a toolbar with playback controls
-     */
-    createUI() {
-        const onRestart = () => {
-            this.video.currentTime = 0;
-        };
-
-        const onSkip = (val) => {
-            this.video.currentTime += val;
-        };
-
-        const onPlayPause = () => {
-            const paused = this.video.paused;
-            console.log(paused);
-
-            if (paused) {
-                this.video.play();
-            } else {
-                this.video.pause();
-            }
-
-            const label = paused ? "||" : "►";
-            this.ui.updateElement("pause", label);
-        };
-
-        const colors = {
-            blue: {
-                light: "#1bf",
-                lighter: "#3df",
-            },
-            red: "#f00",
-            white: "#fff",
-            yellow: {
-                bright: "#ff0",
-                dark: "#bb0",
-            },
-        };
-
-        const config = {
-            panelSize: { width: 2, height: 0.5 },
-            opacity: 1,
-            height: 128,
-            prev: {
-                type: "button",
-                position: { top: 32, left: 0 },
-                width: 64,
-                fontColor: colors.yellow.dark,
-                hover: colors.yellow.bright,
-                onSelect: () => onSkip(-5),
-            },
-            pause: {
-                type: "button",
-                position: { top: 35, left: 64 },
-                width: 128,
-                height: 52,
-                fontColor: colors.white,
-                backgroundColor: colors.red,
-                hover: colors.yellow.bright,
-                onSelect: onPlayPause,
-            },
-            next: {
-                type: "button",
-                position: { top: 32, left: 192 },
-                width: 64,
-                fontColor: colors.yellow.dark,
-                hover: colors.yellow.bright,
-                onSelect: () => onSkip(5),
-            },
-            restart: {
-                type: "button",
-                position: { top: 35, right: 10 },
-                width: 200,
-                height: 52,
-                fontColor: colors.white,
-                backgroundColor: colors.blue.light,
-                hover: colors.blue.lighter,
-                onSelect: onRestart,
-            },
-            renderer: this.renderer,
-        };
-
-        const content = {
-            prev: "<path>M 10 32 L 54 10 L 54 54 Z</path>",
-            pause: "||",
-            next: "<path>M 54 32 L 10 10 L 10 54 Z</path>",
-            restart: "Restart",
-        };
-
-        const ui = new CanvasUI(content, config);
-        ui.mesh.position.set(0, -1, -3);
-
-        return ui;
-    }
-
-    createProgressBar() {
-        const barGroup = new THREE.Group();
-
-        const bgBarGeometry = new THREE.PlaneGeometry(1, 0.1);
-        const bgBarMaterial = new THREE.MeshBasicMaterial({ color: 0xffffff });
-        const bgBarMesh = new THREE.Mesh(bgBarGeometry, bgBarMaterial);
-
-        const barGeometry = new THREE.PlaneGeometry(1, 0.1);
-        const barMaterial = new THREE.MeshBasicMaterial({ color: 0xff031b });
-        const barMesh = new THREE.Mesh(barGeometry, barMaterial);
-
-        const { x, y, z } = this.ui.mesh.position;
-        barMesh.name = "red progress bar";
-        barGroup.add(barMesh);
-        bgBarMesh.name = "white progress bar";
-        barGroup.add(bgBarMesh);
-        barGroup.position.set(x, y + 0.3, z);
-
-        return barGroup;
-    }
-
-    updateProgressBar() {
-        const redProgressBar = this.progressBar.getObjectByName(
-            "red progress bar"
-        );
-        const whiteProgressBar = this.progressBar.getObjectByName(
-            "white progress bar"
-        );
-
-        const progress = (this.video.currentTime / this.video.duration) * 2;
-        redProgressBar.scale.set(progress, 1, 1);
-        const redOffset = (2 - progress) / 2;
-        redProgressBar.position.x = -redOffset;
-        redProgressBar.position.needsUpdate = true;
-
-        whiteProgressBar.scale.set(2 - progress, 1, 1);
-        const whiteOffset = progress / 2;
-        whiteProgressBar.position.x = whiteOffset;
-        whiteProgressBar.position.needsUpdate = true;
-    }
-
-    setVideoCurrentTime(xPosition) {
-        // Set video playback position
-        const timeFraction = (xPosition + 1) / 2;
-        this.video.currentTime = timeFraction * this.video.duration;
-    }
-
-    createToolbarGroup() {
-        const toolbarGroup = new THREE.Group();
-        toolbarGroup.add(this.ui.mesh);
-        toolbarGroup.add(this.progressBar);
-
-        toolbarGroup.position.set(0, 1.6, -2);
-        toolbarGroup.rotateX(-Math.PI / 4);
-
-        return toolbarGroup;
-    }
-
-    /**
      * Creates an HTML video using `videoIn` as src attribute
      * @param {} videoIn video.src
      */
@@ -392,24 +235,16 @@ class App {
                 .set(0, 0, -1)
                 .applyMatrix4(worldMatrix);
 
-            const intersections = this.raycaster.intersectObjects([
-                this.ui.mesh,
-                ...this.progressBar.children,
-            ]);
+            const intersections = this.raycaster.intersectObjects(
+                this.toolbar.objects
+            );
 
             if (intersections.length === 0) {
                 this.scene.userData.isToolbarVisible = false;
                 this.scene.remove(this.toolbarGroup);
             }
 
-            const intersectionWithProgressBar = intersections.find(
-                ({ object: { name } }) =>
-                    name === "white progress bar" || name === "red progress bar"
-            );
-            if (intersectionWithProgressBar) {
-                this.setVideoCurrentTime(intersectionWithProgressBar.point.x);
-                this.updateProgressBar();
-            }
+            this.toolbar.update(intersections);
         }
     }
